requirements.txt
langchain
python-dotenv
wryte[color]
colorama
google-api-python-client